changes for metrics

+++ business/strategy/ping_pong.go
@@ -19,7 +19,7 @@ func HandlePingPong(wsconn *websocket.Conn, testState *models.TestResult, queue 
 			if testState.ReceivedMsgCount == 1{
 				testState.ReceiveTimeFirst = time.Now()
 			}
-			testState.Latency = testState.Latency - float64(testState.SendTimeLatest.UnixMilli())
+			testState.Latency = testState.Latency - testState.SendTimeLatest.UnixMilli()
 			testState.EventType = "receive"
 			testState.EventTime = time.Now()
 			if err != nil {
@@ -41,7 +41,7 @@ func HandlePingPong(wsconn *websocket.Conn, testState *models.TestResult, queue 
 			testState.SendTimeFirst =  time.Now()
 		}
 		testState.EventType = "send"
-		testState.Latency = testState.Latency - float64(testState.SendTimeLatest.UnixMilli())
+		testState.Latency = testState.Latency - testState.SendTimeLatest.UnixMilli()
 		testState.EventTime = time.Now()
 		wsconn.WriteMessage(websocket.TextMessage, []byte("ping"))
 		log.Info(nil).Str("message", string("ping")).Msg("sent message to ws")
